% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mimic_preprocessed}
\alias{mimic_preprocessed}
\title{Preprocessed Clinical Data from the MIMIC-II Database}
\format{
A data frame with fewer variables than the original due to preprocessing. Number of rows: 1776.
}
\source{
\url{https://physionet.org/content/mimic2-iaccd/1.0/}
}
\usage{
mimic_preprocessed
}
\description{
This version of the \code{mimic} dataset has been cleaned by removing columns with more than 10\% missing data, imputing remaining missing values with the median, and dropping columns highly correlated with the outcome. It is designed for use in fairness-aware machine learning tasks and streamlined analysis.
}
\seealso{
\code{\link{mimic}}
}
\keyword{datasets}
